Reorganise helpers and constants
