+++ src/content/app/genome-browser/components/browser-nav/BrowserNavConfig.ts
