+++ src/content/app/genome-browser/constants/browser-constants.ts
